1) Corrección de bug en el reproductor de audio 2) Nuevo cuento para GAME6

+++ app/src/main/java/csim/csimemotions/Generic_Game.java
@@ -40,7 +40,7 @@ public abstract class Generic_Game extends android.support.v4.app.Fragment imple
     protected LogManager logMan;
     protected Log logSession;
     private LogStage logStage;
-    private byte gamePoints;
+    private int gamePoints;
 
     protected int stageNumber;
     private boolean wasSoundPlaying;
